# internal/platform/pg/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package queries

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Acknowledgement struct {
	Fingerprint    string
	StartsAt       time.Time
	AcknowledgedBy string
	AcknowledgedAt time.Time
}

type AlertHistory struct {
	ID          int64
	Fingerprint string
	Name        string
	Status      string
	Labels      []byte
	StartsAt    time.Time
	EndsAt      pgtype.Timestamptz
	ObservedAt  time.Time
}

type Session struct {
	ID                 uuid.UUID
	Subject            string
	RefreshTokenSealed []byte
	CreatedAt          time.Time
	RenewedAt          time.Time
	ExpiresAt          time.Time
	Name               string
}

type Silence struct {
	ID          string
	Fingerprint string
	CreatedBy   string
	CreatedAt   time.Time
	EndsAt      pgtype.Timestamptz
	ExpiredAt   pgtype.Timestamptz
}