# openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: estate-dashboard API
  version: 1.0.0
  description: >-
    The contract the Go server (ogen) and the Vue client (@hey-api/openapi-ts) are both generated
    from. Change this document first; `task gen` regenerates both sides and `task gen:check` fails
    when either is stale.
  contact:
    name: Joris Jonkers
    url: https://github.com/JorisJonkers-dev
  license:
    name: Proprietary
    url: https://github.com/JorisJonkers-dev/estate-dashboard/blob/main/LICENSE
servers:
  - url: https://estate.jorisjonkers.dev
    description: Production. The SPA and the API share one origin.
    x-internal: false
security:
  - sessionCookie: []
tags:
  - name: session
    description: Who is signed in.
  - name: alerts
    description: What the dashboard itself recorded about alerts.
paths:
  /api/v1/session:
    get:
      operationId: getSession
      summary: Say who is signed in
      description: >-
        Returns the admin the session cookie belongs to. A request without a session, or with one
        auth no longer stands behind, gets a 401: the web app then shows the sign-in page.
      tags: [session]
      responses:
        "200":
          description: The signed-in admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/alerts/history:
    get:
      operationId: listAlertHistory
      summary: List the states alerts were seen in, newest first
      description: >-
        Returns at most `limit` events, newest first. An event is one state of one firing of an
        alert: it fired, or it resolved. Alertmanager forgets a resolved alert; this is the
        dashboard's own record.
      tags: [alerts]
      parameters:
        - name: limit
          in: query
          required: false
          description: How many events to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: The events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertHistory"
        "400":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-estate_session
      description: >-
        The dashboard's own session, set when an admin signs in through auth at `/auth/login`. The
        cookie holds a sealed session id and nothing else. The server refuses a request without a
        live session of an account that holds the admin role; on a local run DEV_USER stands in.
  responses:
    Problem:
      description: An RFC 9457 problem.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    AlertEvent:
      description: One state an alert was seen in.
      type: object
      additionalProperties: false
      required: [id, fingerprint, name, status, startsAt, observedAt]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          maximum: 9223372036854775807
          description: The event's id.
          example: 42
        fingerprint:
          type: string
          minLength: 1
          maxLength: 64
          description: Alertmanager's fingerprint of the alert.
          example: 9f2c1a7b3d4e5f60
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: The alert's name.
          example: CollectorStopped
        status:
          type: string
          enum: [firing, resolved]
          description: The state the alert was seen in.
          example: resolved
        startsAt:
          type: string
          format: date-time
          maxLength: 64
          description: When this firing of the alert started.
          example: "2026-10-03T07:00:00Z"
        endsAt:
          type: string
          format: date-time
          maxLength: 64
          description: When the alert resolved. Present exactly when the status is resolved.
          example: "2026-10-03T07:30:00Z"
        observedAt:
          type: string
          format: date-time
          maxLength: 64
          description: When the dashboard saw the alert in this state.
          example: "2026-10-03T07:30:20Z"
    Session:
      description: The signed-in admin.
      type: object
      additionalProperties: false
      required: [subject, name]
      properties:
        subject:
          type: string
          minLength: 1
          maxLength: 255
          description: The account's id in auth.
          example: 3f6c2a1e-8b7d-4c5f-9a10-2b3c4d5e6f70
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: What to call the admin, which is the account's username.
          example: joris
    AlertHistory:
      description: A page of events, newest first.
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          description: The events.
          items:
            $ref: "#/components/schemas/AlertEvent"
    Problem:
      description: An RFC 9457 problem detail.
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          maxLength: 256
          description: A URI naming the problem type.
          example: about:blank
        title:
          type: string
          maxLength: 256
          description: A short summary of the problem type.
          example: Bad request
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
          example: 400
        detail:
          type: string
          maxLength: 1024
          description: What went wrong in this occurrence.
          example: The limit is out of range.